requirements.txt
streamlit==1.24.1
python-dotenv==1.0.0
langchain==0.0.200
langchain-google-genai==0.0.6
langchain-community==0.0.39
langchain-core==0.0.6
faiss-cpu==1.7.4
google-api-core==2.11.0
google-auth==2.22.0
protobuf==4.23.4
grpcio==1.56.0
unstructured[local-inference]==0.7.9